// src/app/models/application-tracking-detail.ts
export interface ApplicationTrackingDetail {
  // clarification required state
  clarificationRequired: boolean;
  // prior approval completed state
  priorApprovalCompleted: boolean;
  // post implementation ready state
  postImplementationReady: boolean;
  // post implementation submitted state
  postImplementationSubmitted: boolean;
}